[layout] moved the font styles to a layout file for home route
